requirements.txt
torchvision==0.4.1a0+d94043a
torch==1.3.0
Pillow==7.2.0
